refactor: functions
