+++ src/views/DesignEditor/components/Toolbox/Shared/LockUnlock.tsx
@@ -3,8 +3,7 @@ import { useActiveObject, useEditor } from "@layerhub-io/react"
 import Locked from "~/components/Icons/Locked"
 import Unlocked from "~/components/Icons/Unlocked"
 import { ToolButton } from "./ToolButton"
-import { LockFunc } from "~/views/DesignEditor/utils/tools/LockFunc"
-import { UnlockFunc } from "~/views/DesignEditor/utils/tools/UnlockFunc"
+import { LockFunc, UnlockFunc } from "~/views/DesignEditor/utils/tools/LockUnlockFunc"
 
 const LockUnlock = ({ type }: any) => {
   const [state, setState] = React.useState<{ locked: boolean }>({ locked: false })
