+++ src/views/DesignEditor/utils/tools/GroupUngroupFunc.tsx
@@ -3,3 +3,9 @@ export const GroupFunc = ({ editor, activeObject, state, setState }: any) => {
   editor.objects.group()
   setState({ ...state, isGroup: true })
 }
+
+export const UngroupFunc = ({ editor, state, setState }: any) => {
+  editor.objects.ungroup()
+  setState({ ...state, isGroup: false })
+  editor.cancelContextMenuRequest()
+}
