--- src/views/DesignEditor/utils/tools/LockFunc.tsx
@@ -1,4 +1,0 @@
-export const LockFunc = ({ editor, setState }: any) => {
-  editor.objects.lock()
-  setState({ locked: true })
-}
