+++ src/views/DesignEditor/utils/tools/LockUnlockFunc.tsx
@@ -1,0 +1,11 @@
+export const LockFunc = ({ editor, setState }: any) => {
+  editor.objects.lock()
+  setState({ locked: true })
+  editor.cancelContextMenuRequest()
+}
+
+export const UnlockFunc = ({ editor, setState }: any) => {
+  editor.objects.unlock()
+  editor.cancelContextMenuRequest()
+  setState({ locked: false })
+}
