--- src/views/DesignEditor/utils/tools/UngroupFunc.tsx
@@ -1,4 +1,0 @@
-export const UngroupFunc = ({ editor, state, setState }: any) => {
-  editor.objects.ungroup()
-  setState({ ...state, isGroup: false })
-}
