--- src/views/DesignEditor/utils/tools/UnlockFunc.tsx
@@ -1,4 +1,0 @@
-export const UnlockFunc = ({ editor, setState }: any) => {
-  editor.objects.unlock()
-  setState({ locked: false })
-}
